Squashed commit of the following:

commit d3126cc352814de29fcdd3feec46049ef098cbc5
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Fri Oct 18 14:01:07 2019 -0700

    fix extended precision warnings from double functions

commit 27c621b66f3d9c5502ddc98ec110d435d0c4179b
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Fri Oct 18 13:06:12 2019 -0700

    remove missed macros

commit 104b24e93bfdee801c36755617cb0c9aad0bcd79
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Fri Oct 18 12:58:22 2019 -0700

    remove unused includes

commit 4c3e14c5b19b9d9295e490f9158acd76efd01769
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Fri Oct 18 12:41:33 2019 -0700

    remove new SUNR macros

commit 4a7bba54acad9b629aaca654d7c3b0342eb82b8b
Merge: e7058bf20 6361e33f1
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Fri Oct 18 10:45:56 2019 -0700

    Merge branch 'develop' into maintenance/fix-conversion-warnings

    Conflicts:
     include/sunlinsol/sunlinsol_pcg.h
     include/sunlinsol/sunlinsol_spbcgs.h
     include/sunlinsol/sunlinsol_spfgmr.h
     include/sunlinsol/sunlinsol_spgmr.h
     include/sunlinsol/sunlinsol_sptfqmr.h
     include/sunlinsol/sunlinsol_superlumt.h

commit e7058bf201a9fddb2e9e76368f904986bc183508
Merge: a08e68c67 53e535e9b
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Fri Oct 18 09:33:00 2019 -0700

    Merge branch 'develop' into maintenance/fix-conversion-warnings

commit a08e68c67a3638f32cfcc5a727df7d0f3cbf0380
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Fri Oct 18 07:50:43 2019 -0700

    fix typo

commit 2894e5dc1d3f788b51a0c1a80e13faad5c86f82d
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Fri Oct 18 07:47:31 2019 -0700

    add note on precision independent macros

commit c0f75ff50c0dc4c33d3db63fd3894a5ff86fb7f0
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Thu Oct 17 20:37:11 2019 -0700

    link example to c math library if GENERIC_MATH is OFF

commit d57fca975c78d69f71b27ca3b3d16e4fadb6c9a8
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Thu Oct 17 19:00:51 2019 -0700

    set USE_GENERIC_MATH in test script based on c std

commit 4cf06c3c8f43f37442b0df6959041a10ed1f9573
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Thu Oct 17 18:45:51 2019 -0700

    fix minor petsc conversion warnings

commit fdf6bd4ab69b92350c2f38c3ecb2aebbdc89646c
Merge: 9de179fc7 56792909d
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Thu Oct 17 18:11:11 2019 -0700

    Merge branch 'develop' into maintenance/fix-conversion-warnings

commit 9de179fc7ea32739668c9ce4ad2a0535e69c695c
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Thu Oct 17 16:09:04 2019 -0700

    use cmath rather than math.h in C++ examples

commit fb35d050ff941534f3a8fda63059fd98a00c770f
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Thu Oct 17 15:09:11 2019 -0700

    remove old mpicuda, mpiraja headers

commit c9b448daefe537cb609f8ad92ee2af05e7c4b41c
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Thu Oct 17 15:06:59 2019 -0700

    remove unsupported fnvector_parhyp

commit b00342a455e8285241a85aabbe363f7dd3068457
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Thu Oct 17 15:03:52 2019 -0700

    remove unnecessary include

commit 438d6541ed71511a4de367aca1b95c0889903015
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Thu Oct 17 06:27:40 2019 -0700

    revert some cherry-pick changes

commit b078bdfcee0e2abcffe31ff90a647dd8d1e20b2d
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Thu Oct 17 05:41:46 2019 -0700

    fix conversion warning in trilinos vector

commit f2420c97fc20ad29105bc031dc1c16ad685a264d
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Wed Oct 16 18:57:44 2019 -0700

    fix merge errors

commit e3cf1a3d52534685297207786c3e40272a697c2b
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Sun Oct 13 19:16:24 2019 -0700

    add -Wconversion -Wno-sign-conversion for some test setups

commit 1aebe90195288d951887b059318c322b3215ad64
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Tue Oct 15 11:07:20 2019 -0700

    fix conversion warning

commit fd880a0f929f549fc4e3f5abeed1975e915f67f4
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Sun Oct 13 17:35:00 2019 -0700

    fix F77 interface long int conversion warnings

commit 5bad3a780d14528187cb444b87e38015e237ffdb
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Sun Oct 13 12:59:27 2019 -0700

    fix int64 conversion warnings

commit 9bcece0698e81638be151e48db1452dc85300275
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Sat Oct 12 15:22:10 2019 -0700

    fix extended precision conversion warnings

commit e4cc10dfd0c8caa8dcc3a66a151547a52745cb51
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Sat Oct 12 14:23:05 2019 -0700

    fix SOME more conversion warnings in example programs

commit 117ce8d06568018c41e0a38579c857c05df7cd71
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Fri Oct 11 22:49:44 2019 -0700

    fix SOME conversion warnings in example programs

commit 3c55e30bf68410848566cedfe2aac0682f405e01
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Fri Oct 11 22:17:32 2019 -0700

    fix comment alignment

commit cba0b013bfb015bd810732e9b9a57740d358051e
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Fri Oct 11 17:57:41 2019 -0700

    whitespace cleanup

commit 88562a1cd098896ef5e3a9800864d5c3707d7762
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Fri Oct 11 17:07:01 2019 -0700

    fix type conversion warnings in packages

+++ examples/arkode/C_serial/ark_brusselator_fp.c
@@ -85,7 +85,7 @@ int main()
   realtype Tf = RCONST(10.0);    /* final time */
   realtype dTout = RCONST(1.0);  /* time between outputs */
   sunindextype NEQ = 3;          /* number of dependent vars. */
-  int Nt = ceil(Tf/dTout);       /* number of output times */
+  int Nt = (int) ceil(Tf/dTout); /* number of output times */
   int test = 3;                  /* test problem to run */
   realtype reltol = 1.0e-6;      /* tolerances */
   realtype abstol = 1.0e-10;
